Moved the old access code code to the new discount codes controller
